Created file CLIUtilities.py. Moved DirectoryListing from fcCLI into this new file.
